Fix memory-safety and signature defects found by the Phase 8 code review
A four-axis review pass over the 20 Phase 8 commits. This fixes the defects that
are memory-unsafe or corrupt native state; the leak and collection findings
follow separately.

P/Invoke signatures (would corrupt memory, not just return wrong values):

- cna_dual_texture_effect_get/set_texture were declared without their
  texture_index parameter, so every later argument sat one slot left: set passed
  the texture handle where native reads an index and read garbage as the handle.
  Fixing it also makes DualTextureEffect.Texture2 real -- the parameter IS the
  second-layer route ("Texture layer index, zero or one"), so the property that
  threw NotSupportedException now works. The earlier claim that the C API had no
  second-layer function came from the missing parameter, not from the API.
- cna_graphics_adapter_get_display_mode_count/_copy_display_modes were declared
  without filter_by_format and format, shifting out_count and the destination
  pointer two slots -- native would have written a uint64 and a mode array
  through stale registers.
- cna_graphics_adapter_get_current_display_mode and
  cna_graphics_device_get_display_mode used `out` for a struct the header
  documents as caller-initialized; `out` skips the parameterless constructor, so
  struct_size/struct_version arrived zero. Now `ref` over a constructed local,
  matching every other versioned-struct call here.
- cna_game_components_index_of declared out long for an int32_t out-param.

Memory safety:

- VideoPlayer.GetTexture wrapped a handle video.h documents as BORROWED ("valid
  only until the next call on this player") in an owning SafeHandle. Its critical
  finalizer would have destroyed the player's own frame texture on the next GC,
  whether or not anyone disposed it -- a doc comment saying "do not dispose this"
  cannot prevent that. NativeResourceHandle/Texture gained a non-owning mode.
- GameComponent.TryResolve ran outside every callback's try block, and
  GCHandle.Target throws on a freed handle -- so a stale callback unwound an
  exception into native code (CoreCLR fails fast). It is now total: a stale
  handle makes the callback a no-op, which is the only safe answer there.
- The native dispose callback called Dispose(true), which calls
  cna_game_component_destroy -- but the header documents that handler as running
  "before its handle is released". Managed code was releasing a handle native was
  about to use. It now does managed-side teardown only.

Correctness:

- StockEffect gained a ReleaseAdditionalNativeResources hook so subclass cleanup
  runs INSIDE the _disposed guard. EnvironmentMapEffect and SkinnedEffect
  overrode Dispose instead, putting three cna_directional_light_destroy calls
  before the guard -- a double free on any second Dispose(). BasicEffect had it
  right; the two newer effects had not copied the guard. The hook makes it
  unrepresentable rather than something each subclass must remember.
- EffectMaterial required its clone source to be a StockEffect, which the compat
  effects are not (they report their handle by overriding NativeEffectHandleValue)
  -- so cloning any compat effect threw, with a message wrongly claiming it had no
  native effect. It now reads the handle through Effect's own accessor.
- GameComponent.ThrowPendingException rethrew a NEW wrapper each time. Since it is
  reachable from inside a callback, the wrapper was re-caught and re-wrapped,
  growing one InnerException layer per frame without bound. It now rethrows the
  original and puts the explanation in Data.
- A captured exception was discarded forever if the component never read a guarded
  property; Dispose now surfaces it, and later failures are counted and reported.
- Enabled/UpdateOrder raised their change events on no-op writes;
  runtime_components.h says setting a value to what it already is must not.

Build 0 warnings/0 errors; 659 tests passing.

+++ src/CNA.Framework/GameComponentCollection.cs
@@ -104,9 +104,9 @@ public class GameComponentCollection : ICollection<GameComponent>
     {
         ArgumentNullException.ThrowIfNull(item);
 
-        CnaResult result = Native.cna_game_components_index_of(GameHandle, item.NativeHandle, out long index);
+        CnaResult result = Native.cna_game_components_index_of(GameHandle, item.NativeHandle, out int index);
         CnaException.ThrowIfFailed(result, nameof(IndexOf));
-        return (int)index;
+        return index;
     }
 
     public void CopyTo(GameComponent[] array, int arrayIndex)
